Adding http lib
I decided to just add the HTTP lib to the repo.  I also made it so
maven will pick up the lib using the pom.  This way, the library is
more portable and takes less time to set up and build.  I also
made one of the unit tests not fail when insufficient funds are
available to create an order.